Request an authentication token only at join, and using the GETTOKEN action instead of the LIST action. (#351)
Requesting an authentication token no longer fetches the entire server list. The GETTOKEN action is used to fetch a token when connecting to a server, and the server host/port are passed along. This also instructs the list server to not generate a token for the LIST action, since we're now trying to restrict tokens to a specific server instead of the player IP in order to  handle CGNAT.