Work around the GLEW issue on Wayland
When running the client using SDL's Wayland driver, glewInit() was
returning GLEW_ERROR_NO_GLX_DISPLAY. This was messing up our OpenGL
state because we were bailing out of initContext(). We don't use any GLX
extensions, so it seems like it's safe for us to just ignore this and
continue.

Fixes #377